[Solve] 22.02.18. 5주차 DP 문제 풀이
1904 01타일
